refactor(analyzer)!: Move `PackageCurationProvider` to model
Not all implementations can be moved due to dependencies on the
downloader module. So, move only the interface for now to prepare for an
upcoming change.
